requirements.txt
Django==2.2
argparse
gunicorn
Pillow
django-import-export